Create plotdiff
This connects to an Excel file which has around 6000 data points for each of 3 reference points, the x axis, and one sample. The program would calculate the absolute difference of each point in the sample column to all combinations of the references and tell you which one it was closest to. It then constructs a plot.